requirements.txt
torch>=2.2.0          # core tensor & GPU support
transformers>=4.41.0  # HF models / tokenizers
datasets>=2.19.0      # HF datasets loader
peft>=0.10.0          # LoRA fine-tuning
accelerate>=0.29.0    # multi-GPU launcher
deepspeed>=0.14.0     # ZeRO-3 sharded training
bitsandbytes>=0.43.0  # 4-bit weight loading (optional)
evaluate>=0.4.1       # task metrics (SQuAD, accuracy)
loguru>=0.7.2         # nicer logging
hydra-core>=1.3.2     # config management
omegaconf>=2.3.0      # Hydra dependency
tqdm>=4.66.4          # progress bars

